Totally replace the camera2 implementation
