+++ sphinx/screens-detail/camera/camera/src/main/java/chat/sphinx/camera/navigation/CameraNavigator.kt
@@ -10,19 +10,8 @@ abstract class CameraNavigator(
 ): Navigator<NavController>(navigationDriver) {
 
     @JvmSynthetic
-    internal suspend fun toCameraDetailScreen(
-        replacingVideoFragment: Boolean = false
-    ) {
-        navigationDriver.submitNavigationRequest(
-            ToCameraDetail(
-                replacingVideoFragment = replacingVideoFragment
-            )
-        )
-    }
-
-    @JvmSynthetic
-    internal suspend fun toCaptureVideoDetailScreen() {
-        navigationDriver.submitNavigationRequest(ToCaptureVideoDetail())
+    internal suspend fun toCameraDetailScreen() {
+        navigationDriver.submitNavigationRequest(ToCameraDetail())
     }
 
     @JvmSynthetic
