--- sphinx/screens-detail/camera/camera/src/main/java/chat/sphinx/camera/ui/CaptureVideoFragment.kt
@@ -1,523 +1,0 @@
-package chat.sphinx.camera.ui
-
-import android.Manifest
-import android.annotation.SuppressLint
-import android.content.Context
-import android.content.pm.PackageManager
-import android.hardware.camera2.*
-import android.os.Bundle
-import android.view.*
-import android.widget.ImageView
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.camera.core.*
-import androidx.camera.lifecycle.ProcessCameraProvider
-import androidx.camera.video.*
-import androidx.camera.video.VideoCapture
-import androidx.concurrent.futures.await
-import androidx.core.content.ContextCompat
-import androidx.core.net.toFile
-import androidx.core.net.toUri
-import androidx.core.util.Consumer
-import androidx.fragment.app.FragmentActivity
-import androidx.fragment.app.viewModels
-import androidx.lifecycle.Observer
-import androidx.lifecycle.lifecycleScope
-import app.cash.exhaustive.Exhaustive
-import chat.sphinx.camera.R
-import chat.sphinx.camera.databinding.FragmentCaptureVideoBinding
-import chat.sphinx.camera.model.CameraItem
-import chat.sphinx.camera.model.LensFacing
-import chat.sphinx.camera.ui.viewstate.CameraViewState
-import chat.sphinx.camera.ui.viewstate.CapturePreviewViewState
-import chat.sphinx.concept_image_loader.ImageLoader
-import chat.sphinx.insetter_activity.InsetterActivity
-import chat.sphinx.insetter_activity.addNavigationBarPadding
-import com.example.android.camera.utils.OrientationLiveData
-import dagger.hilt.android.AndroidEntryPoint
-import io.matthewnelson.android_feature_screens.ui.sideeffect.SideEffectFragment
-import io.matthewnelson.android_feature_screens.util.gone
-import io.matthewnelson.android_feature_screens.util.visible
-import io.matthewnelson.android_feature_viewmodel.collectViewState
-import io.matthewnelson.android_feature_viewmodel.submitSideEffect
-import io.matthewnelson.android_feature_viewmodel.updateViewState
-import kotlinx.coroutines.*
-import java.io.File
-import java.util.concurrent.ExecutorService
-import java.util.concurrent.Executors
-import javax.inject.Inject
-
-@AndroidEntryPoint
-internal class CaptureVideoFragment: SideEffectFragment<
-        FragmentActivity,
-        CameraSideEffect,
-        CameraViewState,
-        CameraViewModel,
-        FragmentCaptureVideoBinding,
-        >(R.layout.fragment_camera)
-{
-    @Suppress("PrivatePropertyName")
-    private val PERMISSIONS_REQUIRED = arrayOf(
-        Manifest.permission.CAMERA,
-        Manifest.permission.RECORD_AUDIO,
-    )
-
-    private var _binding: FragmentCaptureVideoBinding? = null
-    override val binding: FragmentCaptureVideoBinding
-        get() = _binding!!
-
-    override val viewModel: CameraViewModel by viewModels()
-
-    private lateinit var imageCapture: ImageCapture
-    private lateinit var videoCapture: VideoCapture<Recorder>
-    private var activeRecording: ActiveRecording? = null
-    private lateinit var recordingState: VideoRecordEvent
-
-    private val mainThreadExecutor by lazy { ContextCompat.getMainExecutor(requireContext()) }
-    private lateinit var cameraExecutor: ExecutorService
-
-    private val captureListener = Consumer<VideoRecordEvent> { event ->
-        // cache the recording state
-        if (event !is VideoRecordEvent.Status)
-            recordingState = event
-
-        updateUI(event)
-
-        if (event is VideoRecordEvent.Finalize) {
-            event.outputResults.outputUri.path?.let {
-                lifecycleScope.launch(viewModel.io) {
-                    val file = File(it)
-                    viewModel.updateMediaPreviewViewState(
-                        CapturePreviewViewState.Preview.VideoPreview(file)
-                    )
-                }
-
-            }
-        }
-    }
-
-    @Inject
-    @Suppress("ProtectedInFinal")
-    protected lateinit var imageLoader: ImageLoader<ImageView>
-
-    private val requestPermissionLauncher by lazy(LazyThreadSafetyMode.NONE) {
-        registerForActivityResult(
-            ActivityResultContracts.RequestMultiplePermissions()
-        ) { response ->
-
-            try {
-                for (permission in PERMISSIONS_REQUIRED) {
-                    if (response[permission] != true) {
-                        throw Exception()
-                    }
-                }
-
-                if (currentViewState !is CameraViewState.Active) {
-                    viewModel.updateViewState(
-                        CameraViewState.Active.BackCamera(viewModel.getBackCamera())
-                    )
-                }
-            } catch (e: Exception) {
-                lifecycleScope.launch(viewModel.mainImmediate) {
-                    viewModel.submitSideEffect(
-                        CameraSideEffect.Notify(getString(R.string.camera_permissions_required))
-                    )
-                }
-            }
-        }
-    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        viewModel
-    }
-
-    @Volatile
-    private var orientationLiveData: OrientationLiveData? = null
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentCaptureVideoBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        cameraExecutor = Executors.newSingleThreadExecutor()
-
-        (requireActivity() as InsetterActivity)
-            .addNavigationBarPadding(binding.includeCameraFooter.root)
-            .addNavigationBarPadding(binding.includeCameraMediaPreview.layoutConstraintCameraImagePreviewFooter)
-
-        if (!hasPermissions(requireContext())) {
-            requestPermissionLauncher.launch(PERMISSIONS_REQUIRED)
-        } else {
-            if (currentViewState !is CameraViewState.Active) {
-                viewModel.updateViewState(
-                    CameraViewState.Active.BackCamera(viewModel.getBackCamera())
-                )
-            }
-        }
-
-        binding.includeCameraFooter.imageViewCameraFooterBackFront.setOnClickListener {
-            @Exhaustive
-            when (currentViewState) {
-                is CameraViewState.Active.BackCamera -> {
-                    viewModel.updateViewState(
-                        CameraViewState.Active.FrontCamera(viewModel.getFrontCamera())
-                    )
-                }
-                null,
-                is CameraViewState.Idle,
-                is CameraViewState.Active.FrontCamera -> {
-                    viewModel.updateViewState(
-                        CameraViewState.Active.BackCamera(viewModel.getBackCamera())
-                    )
-                }
-            }
-        }
-
-        binding.includeCameraMediaPreview.apply {
-            imageViewCameraImagePreview.setOnClickListener {
-                viewModel
-            }
-            textViewCameraImagePreviewRetake.setOnClickListener {
-                @Exhaustive
-                when (val vs = viewModel.currentCapturePreviewViewState) {
-                    is CapturePreviewViewState.None -> {}
-                    is CapturePreviewViewState.Preview.ImagePreview -> {
-                        viewModel.deleteImage(vs.media)
-                        viewModel.updateMediaPreviewViewState(CapturePreviewViewState.None)
-                    }
-                    is CapturePreviewViewState.Preview.VideoPreview -> {
-                        viewModel.deleteImage(vs.media)
-                        viewModel.updateMediaPreviewViewState(CapturePreviewViewState.None)
-                    }
-                }
-            }
-            textViewCameraImagePreviewUse.setOnClickListener {
-                @Exhaustive
-                when (val vs = viewModel.currentCapturePreviewViewState) {
-                    is CapturePreviewViewState.None -> {}
-                    is CapturePreviewViewState.Preview -> {
-                        textViewCameraImagePreviewRetake.isEnabled = false
-                        viewModel.processSuccessfulResponse(vs)
-                    }
-                }
-            }
-        }
-
-        binding.includeCameraFooter.textViewCameraFooterSwitch.text = "Take Picture"
-        binding.includeCameraFooter.textViewCameraFooterSwitch.setOnClickListener {
-            viewModel.goToCapturePictureFragment()
-        }
-    }
-
-    /**
-     * UpdateUI according to CameraX VideoRecordEvent type:
-     *   - user starts capture.
-     *   - this app disables all UI selections.
-     *   - this app enables capture run-time UI (pause/resume/stop).
-     *   - user controls recording with run-time UI, eventually tap "stop" to end.
-     *   - this app informs CameraX recording to stop with recording.stop() (or recording.close()).
-     *   - CameraX notify this app that the recording is indeed stopped, with the Finalize event.
-     *   - this app starts VideoViewer fragment to view the captured result.
-     */
-    private fun updateUI(event: VideoRecordEvent) {
-        lifecycleScope.launch(viewModel.mainImmediate) {
-            binding.apply {
-                when (event) {
-                    is VideoRecordEvent.Start -> {
-                        includeCameraFooter.imageViewCameraFooterShutter.gone
-                        includeCameraFooter.imageViewCameraStop.visible
-                    }
-                    is VideoRecordEvent.Finalize-> {
-                        includeCameraFooter.imageViewCameraStop.gone
-                        includeCameraFooter.imageViewCameraFooterShutter.visible
-                    }
-                    else -> {
-                        return@launch
-                    }
-                }
-            }
-
-        }
-    }
-
-    override fun onStop() {
-        super.onStop()
-        try {
-            camera?.close()
-        } catch (e: Throwable) {}
-        camera = null
-    }
-
-    private fun hasPermissions(context: Context) = PERMISSIONS_REQUIRED.all {
-        ContextCompat.checkSelfPermission(context, it) == PackageManager.PERMISSION_GRANTED
-    }
-
-    @Volatile
-    private var camera: CameraDevice? = null
-
-    private fun startCamera(cameraItem: CameraItem) {
-        lifecycleScope.launch(viewModel.mainImmediate) {
-            val cameraProvider = ProcessCameraProvider.getInstance(requireContext()).await()
-
-            val cameraSelector = CameraSelector.Builder()
-                .requireLensFacing(cameraItem.lensFacing.toInt())
-                .build()
-
-            val preview = Preview.Builder().setTargetAspectRatio(AspectRatio.RATIO_16_9)
-                .build().apply {
-                    setSurfaceProvider(binding.previewViewCamera.surfaceProvider)
-                }
-
-            imageCapture = ImageCapture.Builder()
-                .setCaptureMode(ImageCapture.CAPTURE_MODE_MINIMIZE_LATENCY)
-                // We request aspect ratio but no resolution to match preview config, but letting
-                // CameraX optimize for whatever specific resolution best fits our use cases
-                .setTargetAspectRatio(screenAspectRatio())
-                // Set initial target rotation, we will have to call this again if rotation changes
-                // during the lifecycle of this use case
-                .setTargetRotation(binding.previewViewCamera.display.rotation)
-                .build()
-
-            val recorder = Recorder.Builder().build()
-            videoCapture = VideoCapture.withOutput(recorder)
-
-            try {
-                cameraProvider.unbindAll()
-                cameraProvider.bindToLifecycle(
-                    requireParentFragment(),
-                    cameraSelector,
-                    videoCapture,
-                    imageCapture,
-                    preview
-                )
-            } catch (exc: Exception) {
-                resetUIandState()
-            }
-
-            binding.includeCameraFooter.imageViewCameraFooterShutter.setOnClickListener {
-                lifecycleScope.launch(viewModel.io) {
-                    val photoFile = viewModel.createFile("jpg", true)
-
-                    // Setup image capture metadata
-                    val metadata = ImageCapture.Metadata().apply {
-                        // Mirror image when using the front camera
-                        isReversedHorizontal = cameraItem.lensFacing == LensFacing.Front
-                    }
-
-                    // Create output options object which contains file + metadata
-                    val outputOptions = ImageCapture.OutputFileOptions.Builder(photoFile)
-                        .setMetadata(metadata)
-                        .build()
-
-                    // Setup image capture listener which is triggered after photo has been taken
-                    imageCapture.takePicture(
-                        outputOptions, cameraExecutor, object : ImageCapture.OnImageSavedCallback {
-                            override fun onError(exc: ImageCaptureException) {
-                                // TODO: Give user feedback on failure
-                            }
-
-                            override fun onImageSaved(output: ImageCapture.OutputFileResults) {
-                                lifecycleScope.launch(viewModel.io) {
-                                    viewModel.updateMediaPreviewViewState(
-                                        CapturePreviewViewState.Preview.ImagePreview(
-                                            output.savedUri?.toFile() ?: photoFile
-                                        )
-                                    )
-                                    delay(200L)
-                                }
-                            }
-                        })
-                }
-
-            }
-            binding.includeCameraFooter.imageViewCameraFooterShutter.setOnLongClickListener { view ->
-                view.gone
-
-                lifecycleScope.launch(viewModel.io) {
-                    if (activeRecording == null || recordingState is VideoRecordEvent.Finalize) {
-                        binding.includeCameraFooter.imageViewCameraFooterShutter.gone
-                        binding.includeCameraFooter.imageViewCameraStop.visible
-
-                        startRecording()
-                    }
-                }
-
-                return@setOnLongClickListener true
-            }
-
-            binding.includeCameraFooter.imageViewCameraStop.setOnClickListener { view ->
-                view.gone
-                lifecycleScope.launch(viewModel.io) {
-                    if (activeRecording != null && recordingState !is VideoRecordEvent.Finalize) {
-                        activeRecording?.stop()
-                        activeRecording = null
-                        delay(200L)
-                    }
-                }
-            }
-            // re-enable button to switch between back/front camera
-            binding.includeCameraFooter.imageViewCameraFooterBackFront.isEnabled = true
-        }
-    }
-
-    private fun screenAspectRatio(): Int {
-        // TODO: Get window width and height
-//        width: Int, height: Int
-//        val previewRatio = max(width, height).toDouble() / min(width, height)
-//        if (abs(previewRatio - RATIO_4_3_VALUE) <= abs(previewRatio - RATIO_16_9_VALUE)) {
-//            return AspectRatio.RATIO_4_3
-//        }
-        return AspectRatio.RATIO_16_9
-    }
-
-    /**
-     * ResetUI (restart):
-     *    in case binding failed, let's give it another change for re-try. In future cases
-     *    we might fail and user get notified on the status
-     */
-    private fun resetUIandState() {
-        lifecycleScope.launch(viewModel.mainImmediate) {
-            binding.includeCameraFooter.imageViewCameraFooterShutter.visible
-            binding.includeCameraFooter.imageViewCameraStop.gone
-
-            // TODO: Prompt user of reset
-        }
-    }
-
-    /**
-     * Kick start the video recording
-     *   - config Recorder to capture to MediaStoreOutput
-     *   - register RecordEvent Listener
-     *   - apply audio request from user
-     *   - start recording!
-     * After this function, user could start/pause/resume/stop recording and application listens
-     * to VideoRecordEvent for the current recording status.
-     */
-    @SuppressLint("MissingPermission")
-    private fun startRecording() {
-        // create MediaStoreOutputOptions for our recorder: resulting our recording!
-        val fileOutputOptions = FileOutputOptions.Builder(
-            viewModel.createFile("mp4", false)
-        ).build()
-
-        // configure Recorder and Start recording to the mediaStoreOutput.
-        activeRecording = videoCapture.output.prepareRecording(requireActivity(), fileOutputOptions)
-                .withEventListener(
-                    mainThreadExecutor,
-                    captureListener
-                )
-                .withAudioEnabled()
-                .start()
-    }
-
-    override suspend fun onSideEffectCollect(sideEffect: CameraSideEffect) {
-        sideEffect.execute(requireActivity())
-    }
-
-    private val orientationObserver = Observer<Int> { orientation ->
-
-    }
-
-    override suspend fun onViewStateFlowCollect(viewState: CameraViewState) {
-
-        binding.includeCameraFooter.imageViewCameraFooterShutter.setOnClickListener(null)
-        binding.includeCameraFooter.imageViewCameraStop.setOnClickListener(null)
-
-        @Exhaustive
-        when (viewState) {
-            is CameraViewState.Idle -> {}
-            is CameraViewState.Active -> {
-                viewState.cameraItem?.let { item ->
-
-                    // disable button to switch between back/front camera
-                    binding.includeCameraFooter.imageViewCameraFooterBackFront.isEnabled = false
-
-                    try {
-                        orientationLiveData?.removeObserver(orientationObserver)
-                        orientationLiveData = OrientationLiveData(binding.root.context, item.characteristics).apply {
-                            observe(viewLifecycleOwner, orientationObserver)
-                        }
-
-                        startCamera(item)
-                    } catch (e: Exception) {}
-                } // TODO: handle null case with no camera available view
-            }
-        }
-    }
-
-    // have to override here to always push
-    // results to onViewStateFlowCollect w/o
-    // comparing to currentViewState
-    override fun subscribeToViewStateFlow() {
-        onStopSupervisor.scope.launch(viewModel.mainImmediate) {
-            viewModel.collectViewState { viewState ->
-                currentViewState = viewState
-                onViewStateFlowCollect(viewState)
-            }
-        }
-        onStopSupervisor.scope.launch(viewModel.io) {
-            viewModel.collectImagePreviewViewState { viewState ->
-                binding.includeCameraMediaPreview.apply {
-                    lifecycleScope.launch(viewModel.mainImmediate) {
-                        @Exhaustive
-                        when (viewState) {
-                            is CapturePreviewViewState.Preview.VideoPreview -> {
-                                if (viewState.media.exists()) {
-                                    root.visible
-                                    videoViewCameraVideoPreview.visible
-
-                                    val uri = viewState.media.toUri()
-                                    videoViewCameraVideoPreview.setVideoURI(uri)
-                                    videoViewCameraVideoPreview.setOnPreparedListener { mediaPlayer ->
-                                        mediaPlayer.start()
-                                    }
-                                    videoViewCameraVideoPreview.setOnErrorListener { _, _, _ ->
-                                        return@setOnErrorListener true
-                                    }
-                                } else {
-                                    viewModel.updateMediaPreviewViewState(CapturePreviewViewState.None)
-                                }
-                            }
-                            is CapturePreviewViewState.Preview.ImagePreview -> {
-                                if (viewState.media.exists()) {
-                                    root.visible
-                                    imageViewCameraImagePreview.visible
-                                    imageLoader.load(imageViewCameraImagePreview, viewState.media)
-                                } else {
-                                    viewModel.updateMediaPreviewViewState(CapturePreviewViewState.None)
-                                }
-                            }
-                            is CapturePreviewViewState.None -> {
-                                root.gone
-                                imageViewCameraImagePreview.setImageDrawable(null)
-                            }
-                        }
-                    }
-
-                }
-            }
-        }
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-
-        // Shut down our background executor
-        cameraExecutor.shutdown()
-
-        // Unregister the broadcast receivers and listeners
-//        displayManager.unregisterDisplayListener(displayListener)
-    }
-
-    companion object {
-        private const val RATIO_4_3_VALUE = 4.0 / 3.0
-        private const val RATIO_16_9_VALUE = 16.0 / 9.0
-    }
-}
