--- sphinx/screens-detail/camera/camera/src/main/res/layout/fragment_capture_video.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.CameraFragment">
-
-    <androidx.camera.view.PreviewView
-        android:id="@+id/preview_view_camera"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <View
-        android:id="@+id/view_camera_overlay"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        app:layout_constraintBottom_toBottomOf="@+id/preview_view_camera"
-        app:layout_constraintEnd_toEndOf="@+id/preview_view_camera"
-        app:layout_constraintStart_toStartOf="@+id/preview_view_camera"
-        app:layout_constraintTop_toTopOf="@+id/preview_view_camera" />
-
-    <include
-        android:id="@+id/include_camera_header"
-        layout="@layout/layout_camera_header"
-        android:visibility="gone" />
-
-    <include
-        android:id="@+id/include_camera_footer"
-        layout="@layout/layout_camera_footer"
-        android:layout_width="match_parent"
-        android:layout_height="@dimen/camera_footer_height"
-        app:layout_constraintBottom_toBottomOf="parent" />
-
-    <include
-        android:id="@+id/include_camera_media_preview"
-        layout="@layout/layout_camera_media_preview"
-        tools:visibility="gone" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
