+++ sphinx/screens-detail/camera/camera/src/main/res/navigation/camera_nav_graph.xml
@@ -3,17 +3,11 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/camera_nav_graph"
-    app:startDestination="@+id/navigation_capture_video_fragment">
+    app:startDestination="@+id/navigation_camera_fragment">
 
     <fragment
         android:id="@+id/navigation_camera_fragment"
         android:name="chat.sphinx.camera.ui.CameraFragment"
-        tools:layout="@layout/fragment_camera" >
-    </fragment>
-
-    <fragment
-        android:id="@+id/navigation_capture_video_fragment"
-        android:name="chat.sphinx.camera.ui.CaptureVideoFragment"
         tools:layout="@layout/fragment_camera" >
     </fragment>
 </navigation>
