+++ sphinx/screens-detail/camera/camera/src/main/res/values/dimens.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <dimen name="camera_header_height">100dp</dimen>
-    <dimen name="camera_footer_height">150dp</dimen>
+    <dimen name="camera_footer_height">160dp</dimen>
 </resources>
